PD-301 피드 목록 조회 쿼리 수정

+++ src/main/java/com/etplus/vo/FeedVO.java
@@ -15,8 +15,8 @@ public record FeedVO(
     // image
     String imagePath,
 
-    Long commentCount,
-    Long likeCount,
+    int commentCount,
+    int likeCount,
     boolean isLiked,
     boolean isCommented
 ) {
